+++ siprix_voip_sdk_android/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.29
+- Fixed parsing SDP with wrong attributes order, which could cause to lose codecs settings
+- Filter out special characters from the 'brandName'
+- Android Flutter: Added ongoing call notification
+- Android Flutter: Updated notifications using CallStyle/MessagingStyle and separate channels
+- Android Flutter: Redesigned Activity's lock screen visibility (enable only when call present)
+  //2026.03.29
+
 ## 1.0.28
 - Fixed bug with losing mute state when switch between calls or make conference
 - Improved handling RtcpMux/BUNDLE options when ICE enabled